refactored code base and folders
