+++ src/App.jsx
@@ -1,25 +1,25 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { useSelector } from "react-redux";
-import Header from "./components/Header";
-import Sidebar from "./components/Sidebar";
-import Login from "./pages/Login";
-import Signup from "./pages/Signup";
+import Header from "./components/common/Header.jsx";
+import Sidebar from "./components/common/Sidebar.jsx";
+import Login from "./components/user/Login.jsx";
+import Signup from "./components/user/Signup.jsx";
 import Dashboard from "./pages/Dashboard";
-import Profile from "./pages/Profile";
-import CourseGrid from "./pages/courses/CourseGrid.jsx";
+import Profile from "./components/user/Profile.jsx";
+import CourseGrid from "./components/course/CourseGrid.jsx";
 import { themeConfig } from "./themeConfig";
-import CourseHome from "./pages/courses/course/CourseHome.jsx";
-import CourseModules from "./pages/courses/course/CourseModules.jsx";
-import CourseAssignment from "./pages/courses/course/CourseAssignment.jsx";
-import AssignmentSubmission from "./pages/courses/course/AssignmentSubmission.jsx";
-import ModuleDetails from "./pages/courses/course/ModuleDetails";
-import ModulesCreate from "./pages/courses/course/ModulesCreate.jsx";
-import CourseCreate from "./pages/courses/course/CourseCreate.jsx";
-import CourseEdit from "./pages/courses/course/CourseEdit.jsx";
-import ResourceCreate from "./pages/courses/course/ResourceCreate.jsx";
-import ModulesEdit from "./pages/courses/course/ModulesEdit.jsx";
-import ResourceUpdate from "./pages/courses/course/ResourceUpdate.jsx";
+import CourseHome from "./components/course/CourseHome.jsx";
+import CourseModules from "./components/module/CourseModules.jsx";
+import CourseAssignment from "./components/assignment/CourseAssignment.jsx";
+import AssignmentSubmission from "./components/assignment/AssignmentSubmission.jsx";
+import ModuleDetails from "./components/module/ModuleDetails.jsx";
+import ModulesCreate from "./components/module/ModulesCreate.jsx";
+import CourseCreate from "./components/course/CourseCreate.jsx";
+import CourseEdit from "./components/course/CourseEdit.jsx";
+import ResourceCreate from "./components/resource/ResourceCreate.jsx";
+import ModulesEdit from "./components/module/ModulesEdit.jsx";
+import ResourceUpdate from "./components/resource/ResourceUpdate.jsx";
 
 const LoggedInLayout = ({ children }) => {
   const { theme } = useSelector((state) => state.theme);
