+++ src/components/common/ThemeSwitcher.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useSelector, useDispatch } from "react-redux";
-import { setTheme } from "../store/themeSlice";
-import { themeConfig } from "../themeConfig";
+import { setTheme } from "../../store/themeSlice.js";
+import { themeConfig } from "../../themeConfig.js";
 
 const themes = [
   { name: "light", label: "Light ☀️" },
