+++ src/components/course/CourseCard.jsx
@@ -1,7 +1,7 @@
 import { BookOpen } from "lucide-react";
 import { useNavigate } from "react-router-dom";
 import { useSelector } from "react-redux";
-import { themeConfig } from "../../themeConfig";
+import { themeConfig } from "../../themeConfig.js";
 import { iconColors } from "../../utils/styles.js";
 import React from "react";
