+++ src/components/course/CourseEdit.jsx
@@ -1,13 +1,14 @@
-import React from "react";
+import React, { useEffect } from "react";
 import axios from "axios";
-import { useNavigate } from "react-router-dom";
+import { useNavigate, useParams } from "react-router-dom";
 import { ArrowLeft } from "lucide-react";
 import { useSelector } from "react-redux";
-import { themeConfig } from "../../../themeConfig";
-import { iconColors } from "../../../utils/styles";
+import { themeConfig } from "../../themeConfig.js";
+import { iconColors } from "../../utils/styles.js";
 
-const CourseCreate = () => {
+const CourseEdit = () => {
   const navigate = useNavigate();
+  const { courseId } = useParams();
   const { theme } = useSelector((state) => state.theme);
   const { bg, text, accentBg, border } = themeConfig[theme];
   const [code, setCode] = React.useState("");
@@ -20,30 +21,89 @@ const CourseCreate = () => {
   const [message, setMessage] = React.useState("");
   const [error, setError] = React.useState("");
   const [dateError, setDateError] = React.useState("");
+  const [formErrors, setFormErrors] = React.useState({
+    code: "",
+    title: "",
+    department: "",
+    credits: "",
+  });
 
-  // Date validation function
-  const validateDates = () => {
+  // Fetch course data on component mount
+  useEffect(() => {
+    const fetchCourse = async () => {
+      try {
+        const response = await axios.get(
+          `http://localhost:8080/api/courses/${courseId}`,
+        );
+        const course = response.data;
+        setCode(course.code);
+        setTitle(course.title);
+        setDepartment(course.department);
+        setDescription(course.description);
+        setCredits(course.credits);
+        setStartDate(course.startDate);
+        setEndDate(course.endDate);
+      } catch (err) {
+        setError("Could not fetch course data");
+        console.error(err);
+      }
+    };
+    fetchCourse();
+  }, [courseId]);
+
+  // Form validation function
+  const validateForm = () => {
+    const errors = {};
+    let isValid = true;
+
+    // Course code validation
+    if (code.length !== 6) {
+      errors.code = "Course code must be exactly 6 characters";
+      isValid = false;
+    }
+
+    // Title validation
+    if (title.length < 4) {
+      errors.title = "Title must be at least 4 characters";
+      isValid = false;
+    }
+
+    // Department validation
+    if (department.length < 2) {
+      errors.department = "Department must be at least 2 characters";
+      isValid = false;
+    }
+
+    // Credits validation
+    const creditsNum = parseInt(credits);
+    if (isNaN(creditsNum) || creditsNum < 1 || creditsNum > 30) {
+      errors.credits = "Credits must be between 1 and 30";
+      isValid = false;
+    }
+
+    // Date validation
     if (startDate && endDate && new Date(startDate) >= new Date(endDate)) {
       setDateError("End date must be later than start date");
-      return false;
+      isValid = false;
+    } else {
+      setDateError("");
     }
-    setDateError("");
-    return true;
+
+    setFormErrors(errors);
+    return isValid;
   };
 
   const handleSubmit = async (event) => {
     event.preventDefault();
 
-    // Validate dates before submission
-    if (!validateDates()) {
-      alert(
-        "End date must be later than start date. Please adjust your dates.",
-      );
-      return; // Prevent form submission
+    // Validate form before submission
+    if (!validateForm()) {
+      alert("Please fix the form errors before submitting.");
+      return;
     }
 
     try {
-      await axios.post(`http://localhost:8080/api/courses`, {
+      await axios.put(`http://localhost:8080/api/courses/${courseId}`, {
         code,
         title,
         department,
@@ -52,18 +112,16 @@ const CourseCreate = () => {
         startDate,
         endDate,
       });
-      setCode("");
-      setTitle("");
-      setDepartment("");
-      setDescription("");
-      setCredits("");
-      setStartDate("");
-      setEndDate("");
-      setMessage("Course  created");
+      setMessage("Course updated");
       setError("");
-      setDateError("");
+      setFormErrors({
+        code: "",
+        title: "",
+        department: "",
+        credits: "",
+      });
     } catch (err) {
-      setError("Could not create course");
+      setError("Could not update course");
       setMessage("");
       console.error(err);
     }
@@ -95,7 +153,7 @@ const CourseCreate = () => {
             className={`bg-opacity-20 ${theme === "light" ? "bg-[#9333ea]" : "bg-[#38bdf8]"} p-4 rounded mb-6`}
           >
             <h1 className={`text-center text-2xl font-medium ${text}`}>
-              Create Course
+              Update Course
             </h1>
           </div>
 
@@ -114,8 +172,18 @@ const CourseCreate = () => {
               className={`w-full px-4 py-2 rounded-md border ${theme === "light" ? "bg-light-bg border-light-accent" : "bg-dark-bg border-dark-accent"} ${text} focus:outline-none focus:ring-2 focus:ring-[#9333ea]`}
               required
               value={code}
-              onChange={(e) => setCode(e.target.value)}
+              onChange={(e) => {
+                setCode(e.target.value);
+                setFormErrors({ ...formErrors, code: "" });
+              }}
             />
+            {formErrors.code && (
+              <p
+                className={`text-sm ${theme === "light" ? "text-red-600" : "text-red-400"}`}
+              >
+                {formErrors.code}
+              </p>
+            )}
           </div>
 
           <div className="mb-4">
@@ -133,9 +201,20 @@ const CourseCreate = () => {
               className={`w-full px-4 py-2 rounded-md border ${theme === "light" ? "bg-light-bg border-light-accent" : "bg-dark-bg border-dark-accent"} ${text} focus:outline-none focus:ring-2 focus:ring-[#9333ea]`}
               required
               value={title}
-              onChange={(e) => setTitle(e.target.value)}
+              onChange={(e) => {
+                setTitle(e.target.value);
+                setFormErrors({ ...formErrors, title: "" });
+              }}
             />
+            {formErrors.title && (
+              <p
+                className={`text-sm ${theme === "light" ? "text-red-600" : "text-red-400"}`}
+              >
+                {formErrors.title}
+              </p>
+            )}
           </div>
+
           <div className="mb-4">
             <label
               htmlFor="department"
@@ -151,9 +230,20 @@ const CourseCreate = () => {
               className={`w-full px-4 py-2 rounded-md border ${theme === "light" ? "bg-light-bg border-light-accent" : "bg-dark-bg border-dark-accent"} ${text} focus:outline-none focus:ring-2 focus:ring-[#9333ea]`}
               required
               value={department}
-              onChange={(e) => setDepartment(e.target.value)}
+              onChange={(e) => {
+                setDepartment(e.target.value);
+                setFormErrors({ ...formErrors, department: "" });
+              }}
             />
+            {formErrors.department && (
+              <p
+                className={`text-sm ${theme === "light" ? "text-red-600" : "text-red-400"}`}
+              >
+                {formErrors.department}
+              </p>
+            )}
           </div>
+
           <div className="mb-4">
             <label
               htmlFor="description"
@@ -172,6 +262,7 @@ const CourseCreate = () => {
               onChange={(e) => setDescription(e.target.value)}
             />
           </div>
+
           <div className="mb-4">
             <label
               htmlFor="credits"
@@ -188,9 +279,20 @@ const CourseCreate = () => {
               min="0"
               required
               value={credits}
-              onChange={(e) => setCredits(e.target.value)}
+              onChange={(e) => {
+                setCredits(e.target.value);
+                setFormErrors({ ...formErrors, credits: "" });
+              }}
             />
+            {formErrors.credits && (
+              <p
+                className={`text-sm ${theme === "light" ? "text-red-600" : "text-red-400"}`}
+              >
+                {formErrors.credits}
+              </p>
+            )}
           </div>
+
           <div className="mb-4">
             <label
               htmlFor="startdate"
@@ -208,11 +310,11 @@ const CourseCreate = () => {
               value={startDate}
               onChange={(e) => {
                 setStartDate(e.target.value);
-                // Clear date error when user changes dates
                 if (dateError) setDateError("");
               }}
             />
           </div>
+
           <div className="mb-4">
             <label
               htmlFor="enddate"
@@ -230,7 +332,6 @@ const CourseCreate = () => {
               value={endDate}
               onChange={(e) => {
                 setEndDate(e.target.value);
-                // Clear date error when user changes dates
                 if (dateError) setDateError("");
               }}
             />
@@ -284,4 +385,4 @@ const CourseCreate = () => {
   );
 };
 
-export default CourseCreate;
+export default CourseEdit;
